# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Student Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Branch": {
        "type": "object",
        "required": [
          "name",
          "address"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique branch ID",
            "example": "branch_123"
          },
          "name": {
            "type": "string",
            "description": "Branch name",
            "example": "Downtown Branch"
          },
          "address": {
            "type": "string",
            "description": "Branch address",
            "example": "123 Main Street, Winnipeg"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number",
            "example": "+1-204-555-1234"
          }
        }
      },
      "BranchCreate": {
        "type": "object",
        "required": [
          "name",
          "address"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Branch name",
            "example": "Downtown Branch"
          },
          "address": {
            "type": "string",
            "minLength": 5,
            "maxLength": 100,
            "description": "Branch address",
            "example": "123 Main Street, Winnipeg"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number",
            "example": "+1-204-555-1234"
          }
        }
      },
      "BranchUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Branch name"
          },
          "address": {
            "type": "string",
            "description": "Branch address"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number"
          }
        }
      },
      "Employee": {
        "type": "object",
        "required": [
          "name",
          "position",
          "email",
          "branchId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique employee ID",
            "example": "emp_001"
          },
          "name": {
            "type": "string",
            "description": "Full name of the employee",
            "example": "Jane Doe"
          },
          "position": {
            "type": "string",
            "description": "Job position of the employee",
            "example": "Manager"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Employee's email address",
            "example": "jane.doe@example.com"
          },
          "branchId": {
            "type": "string",
            "description": "ID of the branch the employee belongs to",
            "example": "branch_123"
          },
          "department": {
            "type": "string",
            "description": "Department name",
            "example": "Finance"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number",
            "example": "+1-204-555-9876"
          }
        }
      },
      "EmployeeCreate": {
        "type": "object",
        "required": [
          "name",
          "position",
          "email",
          "branchId"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Full name of the employee"
          },
          "position": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Job position"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Employee's email address"
          },
          "branchId": {
            "type": "string",
            "description": "Branch ID"
          },
          "department": {
            "type": "string",
            "description": "Department name"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number"
          }
        }
      },
      "EmployeeUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the employee"
          },
          "position": {
            "type": "string",
            "description": "Job position"
          },
          "email": {
            "type": "string",
            "description": "Employee's email address"
          },
          "branchId": {
            "type": "string",
            "description": "Branch ID"
          },
          "department": {
            "type": "string",
            "description": "Department name"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/branches": {
      "get": {
        "summary": "Retrieve all branches",
        "tags": [
          "Branches"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all branches"
          }
        }
      },
      "post": {
        "summary": "Create a new branch",
        "tags": [
          "Branches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/branches/{id}": {
      "get": {
        "summary": "Retrieve a branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique branch ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branch details"
          },
          "404": {
            "description": "Branch not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing branch",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique branch ID"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch updated successfully"
          },
          "404": {
            "description": "Branch not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique branch ID"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Branch deleted successfully"
          },
          "404": {
            "description": "Branch not found"
          }
        }
      }
    },
    "/employees": {
      "get": {
        "summary": "Retrieve all employees",
        "tags": [
          "Employees"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all employees"
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "summary": "Retrieve an employee by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique employee ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee details"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing employee",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique employee ID"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an employee by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique employee ID"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Employee deleted successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      }
    },
    "/employees/branch/{branchId}": {
      "get": {
        "summary": "Retrieve employees by branch ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Branch ID to filter employees"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employees for the specified branch"
          }
        }
      }
    },
    "/employees/department/{department}": {
      "get": {
        "summary": "Retrieve employees by department",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "department",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Department name to filter employees"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employees for the specified department"
          }
        }
      }
    }
  },
  "tags": []
}